Developed front end for adding machinery cost, labor cost, material cost to the add operational cost feature
